+++ merino-settings/src/lib.rs
@@ -90,12 +90,6 @@ pub struct Settings {
     /// off, the suggest request object should be logged, but the
     /// search query should be blank.
     pub log_full_request: bool,
-
-    /// Only used for integration tests.
-    // note `#[cfg(test)]` will cause a compile error as the
-    // `merino_test` proc_macro will not be able to find this
-    // field.
-    pub test_changes: Option<Vec<String>>,
 }
 
 /// Settings for the HTTP server.
@@ -183,6 +177,14 @@ pub struct RemoteSettingsGlobalSettings {
     /// - `http://127.0.0.1`
     /// - `https://firefox.settings.services.mozilla.com`
     pub server: String,
+
+    /// Only used for integration tests.
+    /// This field populates the mock returned remote settings collection
+    /// from inside the `merino_test_macro!(|settings| {...}) call.
+    // note `#[cfg(test)]` will cause a compile error as the
+    // `merino_test` proc_macro will not be able to find this
+    // field.
+    pub test_changes: Option<Vec<String>>,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
